pyproject.toml
[project]
name = "agentic-crypto-trading-system"
version = "0.1.0"
description = ""
authors = [
    {name = "Che",email = "cheonu85@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "redis (>=7.3.0,<8.0.0)",
    "asyncio (>=4.0.0,<5.0.0)",
    "crewai (>=1.10.1,<2.0.0)",
    "langgraph (>=1.1.0,<2.0.0)",
    "sentence-transformers (>=5.2.3,<6.0.0)",
    "scikit-learn (>=1.8.0,<2.0.0)",
    "ccxt (>=4.5.42,<5.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "pydantic-settings (>=2.10.1,<2.11.0)",
    "alembic (>=1.18.4,<2.0.0)"
]

[tool.poetry]
packages = [{include = "agentic_crypto_trading_system", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "pytest-mock (>=3.15.1,<4.0.0)"
]
